backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow==10.3.0
